// desktop/src/types.ts
type TMaybe<T> = T | null | undefined
// ...
export type TPlatformHealthCheck = Readonly<{
  healthy: TMaybe<boolean>
  loginRequired: TMaybe<boolean>
  details: TMaybe<string[]>
}>
// ...
export type TGitCredentialHelperData = Readonly<{
  host: string
  path?: string
  username?: string
  password: string
}>